chore: Fixed build issues with useSearchParams
