+++ apps/garden/app/prijava/zaboravljena-zaporka/page.tsx
@@ -1,26 +1,9 @@
-'use client'
-
-import { useState } from 'react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { Button } from "@signalco/ui-primitives/Button"
-import { Input } from "@signalco/ui-primitives/Input"
+import { Suspense } from 'react'
 import { Card, CardHeader, CardTitle, CardContent } from "@signalco/ui-primitives/Card"
-import { Stack } from '@signalco/ui-primitives/Stack'
-import { Typography } from '@signalco/ui-primitives/Typography'
 import { Key } from 'lucide-react'
+import { ForgotPasswordForm } from './ForgotPasswordForm'
 
 export default function ForgotPasswordPage() {
-    const router = useRouter()
-    const searchParams = useSearchParams()
-    const [email, setEmail] = useState(searchParams.get('email') || '')
-
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault()
-        // TODO: Implement forgot password logic here
-        console.log('Forgot password request for:', email)
-        router.push('/prijava/zaboravljena-zaporka/email-poslan')
-    }
-
     return (
         <div className="container flex items-center justify-center min-h-screen">
             <Card className="w-[350px]">
@@ -31,22 +14,9 @@ export default function ForgotPasswordPage() {
                     <CardTitle className='text-center'>Zaboravljena zaporka</CardTitle>
                 </CardHeader>
                 <CardContent>
-                    <form onSubmit={handleSubmit}>
-                        <Stack spacing={3}>
-                            <Typography level='body2' center>Unesi svoj email za promjenu zaporke</Typography>
-                            <Input
-                                id="email"
-                                type="email"
-                                label='Email'
-                                value={email}
-                                onChange={(e) => setEmail(e.target.value)}
-                                required
-                            />
-                            <Button type="submit" variant="soft" fullWidth>
-                                Pošalji email
-                            </Button>
-                        </Stack>
-                    </form>
+                    <Suspense>
+                        <ForgotPasswordForm />
+                    </Suspense>
                 </CardContent>
             </Card>
         </div>
